LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/cross-window/submit-samedocument-sameorigin-expected.txt



FAIL submitting a <form> element that navigates same-document targeting a same-origin window assert_equals: navigationType expected "push" but got "replace"
